requirements.txt
absl-py==1.2.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cicflowmeter==0.1.6
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dill==0.3.6
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.38.0
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
MouseInfo==0.1.3
netifaces==0.11.0
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pdfkit==1.0.0
Pillow==9.3.0
pip==22.3.1
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyshark==0.5.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.2.1
pywhatkit==5.4
requests==2.28.1
requests-oauthlib==1.3.1
reportlab==4.2.5
rsa==4.9
scapy==2.4.3
scikit-learn==1.0.2
scipy==2.4.2
setuptools==47.1.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
textmagic==2.0.3
threadpoolctl==3.1.0
thundersvm==0.3.12
thundersvm-cpu==0.3.3
twilio==7.15.3
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wikipedia==1.4.0
wrapt==1.14.1
zipp==3.8.1

